Create core test case class
The MyTestCase class has the helper functions to start and stop the
adapter with the provided arguments as well as the methods for connecting
to the master and ColumnStore.

The example test case inherits from the base test case class to
demonstrate that the class works.
