+++ maxscale-cdc-adapter/test/tests/Dockerfile
@@ -4,7 +4,7 @@ RUN yum -y install epel-release && \
     yum -y install python34 python34-pip && \
     pip3 install PyMySQL
 
-COPY config.json src/ /tests/
+COPY src/ /tests/
 COPY startCmd.sh /startCmd.sh
 
 CMD ["/startCmd.sh"]
