+++ maxscale-cdc-adapter/test/tests/src/config.json
